requirements.txt
streamlit
google-generativeai
pandas
gspread
oauth2client
tabulate  # ← この行を追加